refactor(auth): isolate Apple provider implementation
